requirements.txt

boto3
python-dotenv
confluent-kafka
psycopg2-binary